Update subdirectory structure
This includes:
- Create api/, ranker/, include/xapian-letor/
- Modify Makefile to include corresponding restructre changes
- Modify configure.ac to include these changes
- Create xapian-letor.h, where xapian-letor.h includes
	api headers in /include/xapian-letor,
	in a way similar to xapian.h
- Update header file paths in questletor.cc
- Add include path to configure.ac
