+++ xapian-letor/api/Makefile.mk
@@ -1,0 +1,13 @@
+EXTRA_DIST +=\
+	api/Makefile
+
+lib_src +=\
+	api/evalmetric.cc\
+	api/featuremanager.cc\
+	api/featurevector.cc\
+	api/letor.cc\
+	api/letor_features.cc\
+	api/letor_internal.cc\
+	api/ranker.cc\
+	api/ranklist.cc
+
