+++ xapian-letor/include/Makefile.mk
@@ -1,4 +1,18 @@
-xapianincludedir = $(incdir)/xapian
+EXTRA_DIST +=\
+		include/Makefile\
+		include/xapian-letor/Makefile
 
-xapianinclude_HEADERS =\
-	include/xapian/letor.h
+xapianletorincludedir = $(incdir)/xapian-letor
+
+inc_HEADERS =\
+	include/xapian-letor.h
+
+xapianletorinclude_HEADERS =\
+		include/xapian-letor/evalmetric.h\
+		include/xapian-letor/featuremanager.h\
+		include/xapian-letor/featurevector.h\
+		include/xapian-letor/letor.h\
+		include/xapian-letor/letor_features.h\
+		include/xapian-letor/letor_internal.h\
+		include/xapian-letor/ranker.h\
+		include/xapian-letor/ranklist.h
