+++ xapian-letor/include/xapian-letor/featuremanager.h
@@ -22,7 +22,7 @@
 #define FEATURE_MANAGER_H
 
 
-#include <xapian/letor.h>
+#include "letor.h"
 #include "letor_features.h"
 #include "featurevector.h"
 #include "ranklist.h"
