+++ xapian-letor/include/xapian-letor/featurevector.h
