+++ xapian-letor/include/xapian-letor/letor.h
