+++ xapian-letor/ranker/Makefile.mk
@@ -1,0 +1,8 @@
+noinst_HEADERS +=\
+	ranker/svmranker.h
+
+EXTRA_DIST +=\
+	ranker/Makefile
+
+lib_src +=\
+	ranker/svmranker.cc
